Rearrange main app structure
