+++ src/Routes.js
@@ -1,13 +1,18 @@
 import React from 'react'
-import { Route, Switch } from 'react-router-dom'
-import Feed from './screens/Feed'
+import { Redirect, Route, Switch } from 'react-router-dom'
+import Home from './screens/Home'
 import Profile from './screens/Profile'
+import ProposalDetail from './screens/ProposalDetail'
 
 export default function Routes() {
   return (
     <Switch>
+      <Redirect exact from="/" to="/home" />
+      <Route path="/home" component={Home} />
       <Route exact path="/profile" component={Profile} />
-      <Route path="/" component={Feed} />
+      <Route exact path="/proposal/:id" component={ProposalDetail} />
+      <Route exact path="/vote/:id" component={ProposalDetail} />
+      <Redirect to="/home" />
     </Switch>
   )
 }
